+++ helpers/aws/index.js
@@ -1,5 +1,7 @@
 var shared = require(__dirname + '/../shared.js');
 var functions = require('./functions.js');
+var api = require('./api.js');
+var api_multipart = require('./api_multipart.js');
 var regRegions = require('./regions.js');
 var govRegions = require('./regions_gov.js');
 var chinaRegions = require('./regions_china.js');
@@ -34,5 +36,7 @@ var helpers = {
 
 for (var s in shared) helpers[s] = shared[s];
 for (var f in functions) helpers[f] = functions[f];
+for (var a in api) helpers[a] = api[a];
+for (var am in api_multipart) helpers[am] = api_multipart[am];
 
 module.exports = helpers;
