+++ plugins/aws/lambda/lambdaTracingEnabled.js
@@ -10,11 +10,30 @@ module.exports = {
     link: 'https://docs.aws.amazon.com/lambda/latest/dg/services-xray.html',
     recommended_action: 'Modify Lambda functions to activate tracing',
     apis: ['Lambda:listFunctions'],
+    settings: {
+        lambda_whitelist: {
+            name: 'Lambda Functions Whitelisted',
+            description: 'A comma-delimited list of known lambda function Function Names that should be whitelisted',
+            regex: '^.{1,255}$',
+            default: 'Aqua-CSPM-Token-Rotator-Function,-CreateCSPMKeyFunction-,-TriggerDiscoveryFunction-,-GenerateVolumeScanningEx-,-GenerateCSPMExternalIdFu-'
+        }
+    },
 
     run: function(cache, settings, callback) {
         var results = [];
         var source = {};
         var regions = helpers.regions(settings);
+
+        var config = {
+            lambda_whitelist: settings.lambda_whitelist || this.settings.lambda_whitelist.default
+        };
+
+        if (config.lambda_whitelist &&
+            config.lambda_whitelist.length) {
+            config.lambda_whitelist = config.lambda_whitelist.split(',');
+        } else {
+            config.lambda_whitelist = [];
+        }
 
         async.each(regions.lambda, function(region, rcb){
             var listFunctions = helpers.addSource(cache, source,
@@ -37,14 +56,29 @@ module.exports = {
                 if (!lambdaFunc.FunctionArn) continue;
                 var resource = lambdaFunc.FunctionArn;
 
-                if (lambdaFunc.TracingConfig &&
-                    lambdaFunc.TracingConfig.Mode &&
-                    lambdaFunc.TracingConfig.Mode.toUpperCase() === 'ACTIVE') {
+                let whitelisted = false;
+                if (config.lambda_whitelist.length) {
+                    config.lambda_whitelist.forEach(whitelist => {
+                        if (lambdaFunc.FunctionName.indexOf(whitelist) > -1) {
+                            whitelisted = true;
+                        }
+                    });
+                }
+
+                if (whitelisted) {
                     helpers.addResult(results, 0,
-                        'Function has active tracing', region, resource);
+                        'The function ' + lambdaFunc.FunctionName + ' is whitelisted.',
+                        region, lambdaFunc.FunctionArn);
                 } else {
-                    helpers.addResult(results, 2,
-                        'Function does not have active tracing', region, resource);
+                    if (lambdaFunc.TracingConfig &&
+                        lambdaFunc.TracingConfig.Mode &&
+                        lambdaFunc.TracingConfig.Mode.toUpperCase() === 'ACTIVE') {
+                        helpers.addResult(results, 0,
+                            'Function has active tracing', region, resource);
+                    } else {
+                        helpers.addResult(results, 2,
+                            'Function does not have active tracing', region, resource);
+                    }
                 }
             }
