+++ plugins/aws/lambda/lambdaVpcConfig.js
@@ -15,7 +15,7 @@ module.exports = {
             name: 'Lambda Functions Whitelisted',
             description: 'A comma-delimited list of known lambda function Function Names that should be whitelisted',
             regex: '^.{1,255}$',
-            default: 'Aqua-CSPM-Token-Rotator-Function'
+            default: 'Aqua-CSPM-Token-Rotator-Function,-CreateCSPMKeyFunction-,-TriggerDiscoveryFunction-,-GenerateVolumeScanningEx-,-GenerateCSPMExternalIdFu-'
         }
     },
 
@@ -55,9 +55,16 @@ module.exports = {
             for (var f in listFunctions.data) {
                 // For resource, attempt to use the endpoint address (more specific) but fallback to the instance identifier
                 var lambdaFunction = listFunctions.data[f];
+                let whitelisted = false;
+                if (config.lambda_whitelist.length) {
+                    config.lambda_whitelist.forEach(whitelist => {
+                        if (lambdaFunction.FunctionName.indexOf(whitelist) > -1) {
+                            whitelisted = true;
+                        }
+                    });
+                }
 
-                if (config.lambda_whitelist.length &&
-                    config.lambda_whitelist.indexOf(lambdaFunction.FunctionName)>-1) {
+                if (whitelisted) {
                     helpers.addResult(results, 0,
                         'The function ' + lambdaFunction.FunctionName + ' is whitelisted.',
                         region, lambdaFunction.FunctionArn);
